chore: rename package from ablil to python-starter-template
