+++ python_starter_template/foo.py
